requirements.txt
# Core LangChain and LangGraph
langchain==0.3.8
langchain-core==0.3.63
langgraph==0.2.60
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.5
langsmith==0.1.147

# Vector Store and Embeddings
langchain-community==0.3.8
sentence-transformers==3.3.1

# LLM Providers
langchain-google-genai==2.0.5
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
google-generativeai==0.8.3
google-auth==2.43.0
openai==1.109.1

# Core Dependencies
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
click==8.3.1
tqdm==4.67.1

# Development and Testing
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==6.0.0
pytest-xdist==3.6.0
pytest-mock==3.14.0
black==25.11.0
mypy==1.18.2

# Visualization
matplotlib==3.10.7
networkx==3.5
ipython==9.7.0
graphviz==0.21
